Refactor Message and CommandHandler classes
The commit includes refactoring of the Message and CommandHandler classes.
The Message class has been moved to a new location and its header file has been updated.
The CommandHandler class has been deleted.
The Channel [fix] class has been updated with a new enum MODE.
The main_test.cpp file has been deleted.
