+++ includes/Channel.hpp
@@ -8,20 +8,6 @@
 
 #include "Client.hpp"
 
-enum MODE {
-  OPERATOR = 1 << 0,                   // 운영자
-  VOICE = 1 << 1,                      // 발언권
-  INVITE_ONLY = 1 << 2,                // 초대 전용
-  MODERATED = 1 << 3,                  // 중재됨
-  NO_EXTERNAL_MESSAGES = 1 << 4,       // 외부 메시지 금지
-  TOPIC_SETTABLE_BY_OP_ONLY = 1 << 5,  // 토픽 보호
-  SECRET = 1 << 6,                     // 비밀 채널
-  PRIVATE = 1 << 7,                    // 개인 채널
-  USER_LIMIT = 1 << 8,                 // 사용자 수 제한
-  BAN_MASK = 1 << 9,                   // 밴 마스크
-  KEY = 1 << 10                        // 채널 키 (비밀번호)
-};
-
 class Channel {
  private:
   std::string _name;
@@ -37,6 +23,20 @@ class Channel {
   std::vector<Client*> _banned;
 
  public:
+  enum MODE {
+    OPERATOR = 1 << 0,                   // 운영자
+    VOICE = 1 << 1,                      // 발언권
+    INVITE_ONLY = 1 << 2,                // 초대 전용
+    MODERATED = 1 << 3,                  // 중재됨
+    NO_EXTERNAL_MESSAGES = 1 << 4,       // 외부 메시지 금지
+    TOPIC_SETTABLE_BY_OP_ONLY = 1 << 5,  // 토픽 보호
+    SECRET = 1 << 6,                     // 비밀 채널
+    PRIVATE = 1 << 7,                    // 개인 채널
+    USER_LIMIT = 1 << 8,                 // 사용자 수 제한
+    BAN_MASK = 1 << 9,                   // 밴 마스크
+    KEY = 1 << 10                        // 채널 키 (비밀번호)
+  };
+
   Channel(const std::string& name, Client* client);
   ~Channel();
