+++ includes/NumericReply.hpp
@@ -1,60 +1,73 @@
 #ifndef NUMERICREPLY_HPP
-# define NUMERICREPLY_HPP
+#define NUMERICREPLY_HPP
 
 #include <iostream>
 #include <string>
-// #include "Server.hpp"
+
+#include "Server.hpp"
+
+class Server;
+
+#define RPL_WELCOME_STR ("001 RPL_WELCOME")
+
+#define ERR_NOTREGISTERED_STR ("451 ERR_NOTREGISTERED")
+#define ERR_NEEDMOREPARAMS_STR ("461 Need more parameters")
+#define ERR_ALREADYREGISTRED_STR ("462 ERR_ALREADYREGISTRED")
+#define ERR_PASSWDMISMATCH_STR ("464 ERR_PASSWDMISMATCH")
+
+#define ERR_MSGTOOLONG 999
+#define ERR_MSGTOOLONG_STR ("999 ERR_MSGTOOLONG")
 
 typedef enum NUMERIC {
-	//ERR
-    ERR_ALREADYREGISTRED = 462,
-    ERR_BADCHANMASK = 476,
-    ERR_BADCHANNELKEY = 475,
-    ERR_BANNEDFROMCHAN = 474,
-    ERR_CANNOTSENDTOCHAN = 404,
-    ERR_CANTKILLSERVER = 483,
-    ERR_CHANOPRIVSNEEDED = 482,
-    ERR_CHANNELISFULL = 471,
-    ERR_ERRONEUSNICKNAME = 432,
-    ERR_INVITEONLYCHAN = 473,
-    ERR_KEYSET = 467,
-    ERR_NEEDMOREPARAMS = 461,
-    ERR_NICKCOLLISION = 436,
-    ERR_NICKNAMEINUSE = 433,
-    ERR_NONICKNAMEGIVEN = 431,
-    ERR_NOOPERHOST = 491,
-    ERR_NORECIPIENT = 411,
-    ERR_NOSUCHCHANNEL = 403,
-    ERR_NOSUCHNICK = 401,
-    ERR_NOSUCHSERVER = 402,
-    ERR_NOTEXTTOSEND = 412,
-    ERR_NOTONCHANNEL = 442,
-    ERR_NOORIGIN = 409,
-    ERR_PASSWDMISMATCH = 464,
-    ERR_TOOMANYTARGETS = 407,
-    ERR_UMODEUNKNOWNFLAG = 501,
-    ERR_USERSDONTMATCH = 502,
-    ERR_WILDTOPLEVEL = 414,
+  // ERR
+  ERR_ALREADYREGISTRED = 462,
+  ERR_BADCHANMASK = 476,
+  ERR_BADCHANNELKEY = 475,
+  ERR_BANNEDFROMCHAN = 474,
+  ERR_CANNOTSENDTOCHAN = 404,
+  ERR_CANTKILLSERVER = 483,
+  ERR_CHANOPRIVSNEEDED = 482,
+  ERR_CHANNELISFULL = 471,
+  ERR_ERRONEUSNICKNAME = 432,
+  ERR_INVITEONLYCHAN = 473,
+  ERR_KEYSET = 467,
+  ERR_NEEDMOREPARAMS = 461,
+  ERR_NICKCOLLISION = 436,
+  ERR_NICKNAMEINUSE = 433,
+  ERR_NONICKNAMEGIVEN = 431,
+  ERR_NOOPERHOST = 491,
+  ERR_NORECIPIENT = 411,
+  ERR_NOSUCHCHANNEL = 403,
+  ERR_NOSUCHNICK = 401,
+  ERR_NOSUCHSERVER = 402,
+  ERR_NOTEXTTOSEND = 412,
+  ERR_NOTONCHANNEL = 442,
+  ERR_NOORIGIN = 409,
+  ERR_PASSWDMISMATCH = 464,
+  ERR_TOOMANYTARGETS = 407,
+  ERR_UMODEUNKNOWNFLAG = 501,
+  ERR_USERSDONTMATCH = 502,
+  ERR_WILDTOPLEVEL = 414,
 
-	//RPL
-	RPL_UMODEIS = 221,
-    RPL_CHANNELMODEIS = 324,
-    RPL_ENDOFBANLIST = 368,
-    RPL_ENDOFNAMES = 366,
-    RPL_INVITING = 341,
-    RPL_LIST = 322,
-    RPL_LISTEND = 323,
-    RPL_LISTSTART = 321,
-    RPL_NAMREPLY = 353,
-    RPL_NOTOPIC = 331,
-    RPL_TOPIC = 332,
-    RPL_YOUREOPER = 381,
+  // RPL
+  RPL_UMODEIS = 221,
+  RPL_CHANNELMODEIS = 324,
+  RPL_ENDOFBANLIST = 368,
+  RPL_ENDOFNAMES = 366,
+  RPL_INVITING = 341,
+  RPL_LIST = 322,
+  RPL_LISTEND = 323,
+  RPL_LISTSTART = 321,
+  RPL_NAMREPLY = 353,
+  RPL_NOTOPIC = 331,
+  RPL_TOPIC = 332,
+  RPL_YOUREOPER = 381
 } e_numeric;
 
 class NumericReply {
-	public:
-		static std::string getNumericReplyERR(e_numeric numeric, Server &serv);
-		static std::string getNumericReplyRPL(e_numeric numeric, Server &serv);
+ public:
+  static std::string getNumericReplyERR(e_numeric numeric, Server &serv);
+  static std::string getNumericReplyRPL(e_numeric numeric, Server &serv);
 };
 
 #endif
